requirements.txt
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.33.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.5.0
pyyaml>=6.0
tqdm>=4.66.0
